+++ src/flask_exts/usercenter/base_user.py
@@ -1,35 +1,7 @@
-class BaseUser:
+from flask_login.mixins import UserMixin
+
+
+class BaseUser(UserMixin):
     @property
     def is_active(self):
-        return True
-
-    @property
-    def is_authenticated(self):
-        return self.is_active
-
-    @property
-    def is_anonymous(self):
-        return False
-
-    def get_id(self):
-        try:
-            return str(self.id)
-        except AttributeError:
-            raise NotImplementedError("No `id` attribute - override `get_id`") from None
-
-    def __eq__(self, other):
-        """
-        Checks the equality of two `User` objects using `get_id`.
-        """
-        if isinstance(other, BaseUser):
-            return self.get_id() == other.get_id()
-        return NotImplemented
-
-    def __ne__(self, other):
-        """
-        Checks the inequality of two `UserMixin` objects using `get_id`.
-        """
-        equal = self.__eq__(other)
-        if equal is NotImplemented:
-            return NotImplemented
-        return not equal
+        return self.actived