+++ tests/email/test_smtp.py
@@ -2,16 +2,14 @@ import pytest
 from flask_exts.email.smtp import SmtpSSL
 
 
-@pytest.mark.skip(reason="password is empty")
-def test_email():
-    host = "smtp.qiye.aliyun.com"
-    port = "465"
-    user = "test@ojso.com"
-    password = ""  # uncomment this line to test with empty password
-    # password = "gTk94GEeNvujr4zm"
+@pytest.mark.skip(reason="skip real email test.")
+def test_email(app):
+    real_email_sender = app.config.get("REAL_EMAIL_SENDER")
+    if not real_email_sender:
+        return
 
     # set to's email
-    to = "test@ojso.com"
+    to = "david.dong.hua@gmail.com"
     content = "This is a test email."
     subject = "This is a test subject."
     data = {
@@ -20,7 +18,7 @@ def test_email():
         "content": content,
     }
 
-    s = SmtpSSL(host, port, user, password)
+    s = SmtpSSL(**real_email_sender)
     r = s.send(data)
     # print(f"send result: {r}")
     assert not r