+++ tests/security/test_email_confirmation.py
@@ -1,4 +1,0 @@
-import pytest
-
-def test_email_confirm(app):
-    pass